backend/lambda-api/requirements.txt
# Cloud IT Service Dashboard - Backend Dependencies

fastapi==0.110.0
uvicorn[standard]==0.27.1
mangum==0.17.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.0
python-multipart==0.0.9
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
python-dotenv==1.0.1
boto3==1.34.0
